This has been a very large commit. The main purpose was to upgrade from discord.js 12 to discord.js 13 in order to avoid bugs from depriciated features.
This came with some challenges. Buttons were removed for now, because
the discord.js buttons framework is stil running with discord.js 13.

I had to rewrite many parts of the music commands, as discord.js/voice
is now a seperate npm library with a completely different workflow.

As of now, the music features seem to be working fine, but I have still
yet to test the spotify side of them. This will come soon.

One very important thing I created was a dynamic
VoiceConnectionWithPlayer type. I wanted to include a an AudioPlayer
object within the VoiceConnection, so I created a class specifically for
this. Through my code, alias are made to VoiceConnection objects casted
to VoiceConnectionWithPlayer instances.

Another I might change is how I check whether the bot is playing
something. Rather than checking to see if the queueMap element for the
guild is defined, I might check whether the VoiceConnectionWithPlayer's
AudioPlayer attribute is outside of the AudioPlayerState.Idle state.

I will try to find bugs quickly, and fix them efficiently. -Lambda